first refactor for tranqui
