+++ ticketatorapp/core/logs.py
@@ -1,6 +1,4 @@
 from __future__ import unicode_literals
-import models
-
 
 def logger(ticket, user, action, destiny):
     log_destiny = (destiny[:190] + '..') if len(str(destiny)) > 75 else destiny
