+++ ticketatorapp/core/models.py
@@ -9,14 +9,12 @@ from django.forms import ModelForm
 # Added imports
 from django.contrib.auth.models import User, Group
 from datetime import datetime
-from core import views_utils as util
-from core import rights
+from core import rights, utils as util
 from django.core.exceptions import ValidationError
 # Extending the Django User's model
 from django.contrib.auth.models import AbstractUser
 
 # Logger class
-from core.views_logs import logger
 
 
 class TimeStampedModelMixin(models.Model):
@@ -57,73 +55,6 @@ class User(AbstractUser):
         super(User, self).save(*args, **kwargs)
 
 
-# => Auth forms
-class UserForm(ModelForm):
-    date_joined = forms.DateField(
-        widget=forms.SelectDateWidget(), input_formats=['%d/%m/%Y'],initial=util.now)
-    password_first = forms.CharField(
-        label='Initial Password', widget=forms.PasswordInput,
-        required=False,initial='')
-    password_check = forms.CharField(
-        label='Password confirmation', widget=forms.PasswordInput,
-        required=False,initial='')
-    is_active = forms.BooleanField(required=False, initial=True)
-
-    # is_superuser = forms.BooleanField(initial=False)
-    # Pass request to query wich user is trying ot modify the object User
-    def __init__(self, *args, **kwargs):
-        self.request = kwargs.pop("request")
-        super(UserForm, self).__init__(*args, **kwargs)
-
-    class Meta:
-        model = User
-        #fields = ['username', 'first_name', 'last_name', 'email', 'is_superuser', 'is_staff',
-        #                    'is_active', 'rssfeed', ]
-        exclude = ['password']
-
-    def clean_password_check(self):
-        password_raw = self.cleaned_data.get('password_first')
-        password_chk = self.cleaned_data.get('password_check')
-        # First password but not second
-        if not password_chk and password_raw:
-            raise forms.ValidationError("You must confirm your password")
-        # First and second but differents
-        elif password_raw != password_chk:
-            raise forms.ValidationError("Your passwords do not match")
-        # Not updating password
-        elif self.instance.pk and password_chk == "" and password_raw == "":
-            pass
-        else:
-            return password_chk
-
-    # Only admin can set is_superuser to TRUE or change it to normal users
-    def clean_is_superuser(self):
-        cleaned_superuser = self.cleaned_data.get('is_superuser')
-        # If we return a simple "HttpResponse" and the cleaned data
-        # we can raise the error for debug purpouses
-        # return HttpResponse(cleaned_superuser)
-        user_obj = self.request.user
-        # If user exists
-        if self.instance.pk:
-            # If modifier is not admin
-            if user_obj.username != 'admin':
-                # Do nothing, return original instance
-                return self.instance.is_superuser
-            else:
-                # If admin, get the submitted value in the form and clean_it
-                self.is_superuser = cleaned_superuser
-                return self.is_superuser
-
-    def clean_last_login(self):
-        return self.instance.last_login
-
-
-class GroupForm(ModelForm):
-    class Meta:
-        model = Group
-        fields = '__all__'
-
-
 # => Companys
 class Company(TimeStampedModelMixin):
     name = models.CharField(max_length=100)
@@ -134,12 +65,6 @@ class Company(TimeStampedModelMixin):
     # address = pending
     def __unicode__(self):
         return self.name
-
-
-class CompanyForm(ModelForm):
-    class Meta:
-        model = Company
-        fields = '__all__'
 
 
 # => Queues (ex Departments)
@@ -153,12 +78,6 @@ class Queue(TimeStampedModelMixin):
 
     def __unicode__(self):
         return self.name
-
-
-class QueueForm(ModelForm):
-    class Meta:
-        model = Queue
-        fields = '__all__'
 
 
 # => Groups
@@ -207,25 +126,6 @@ class Rights(TimeStampedModelMixin):
             super(Rights, self).save(*args, **kwargs)
 
 
-class RightForm(ModelForm):
-    class Meta:
-        model = Rights
-        fields = '__all__'
-
-    # Check at form stage if registry is created or if we can create it
-    def clean_queue_dst(self):
-        detect_function = Rights.detect_rights_exists(
-            Rights(), self.cleaned_data.get('grp_src'),
-            self.cleaned_data.get('queue_dst'))
-        # Check if no pk assigned and if detect_function['status'] is True
-        if not self.instance.pk and detect_function['status']:
-            raise forms.ValidationError(
-                "Rule already created (" + str(detect_function['numbers'][0]) + ") src=>" +
-                str(self.cleaned_data.get('grp_src')) +
-                " dst=>" + str(self.cleaned_data.get('queue_dst')) + "")
-        return self.cleaned_data.get('queue_dst')
-
-
 # => States
 class State(TimeStampedModelMixin):
     name = models.CharField(max_length=30)
@@ -242,10 +142,7 @@ class State(TimeStampedModelMixin):
         return self.name
 
 
-class StateForm(ModelForm):
-    class Meta:
-        model = State
-        fields = '__all__'
+
 
 
 # => Prioritys
@@ -254,12 +151,6 @@ class Priority(TimeStampedModelMixin):
 
     def __unicode__(self):
         return self.name
-
-
-class PriorityForm(ModelForm):
-    class Meta:
-        model = Priority
-        fields = '__all__'
 
 
 # => Inventory Servers/PC
@@ -324,119 +215,6 @@ class Ticket(TimeStampedModelMixin):
             create_user=self.str_creator_user_name(),
             assigned_user_data=self.str_assigned_user_name(),)
 
-
-class TicketForm(ModelForm):
-    #  Pass request to a form => http://stackoverflow.com/questions/6325681/
-    #  passing-a-user-request-to-forms
-    def __init__(self, *args, **kwargs):
-        self.request = kwargs.pop("request")
-        super(TicketForm, self).__init__(*args, **kwargs)
-
-    class Meta:
-        model = Ticket
-        fields = '__all__'
-        # exclude = ['date']
-
-    '''
-    We log all importante changes in fields to be tracked
-    The field_checker test if both are the same and if it found changes, call "logger" passing data
-    TODO: Maybe, "if self.instance.pk is not None:" can be improved to not call
-    all  times to make the check (and save N checks at save time)
-    '''
-
-    def clean_assigned_state(self):
-        if self.instance.pk is not None:   # new instance only
-            self.field_checker(str(self.instance.assigned_state), str(
-                self.cleaned_data.get('assigned_state')))
-        return self.cleaned_data.get('assigned_state')
-
-    '''
-    Check if the new queue assigned is permited...
-    '''
-    def clean_assigned_queue(self):
-        if self.instance.pk is not None:   # new instance only
-            user_object_rights = rights.get_rights_for_ticket(
-                user=self.request.user,
-                queue=self.cleaned_data.get('assigned_queue'), ticket_id=None)
-            if not user_object_rights.can_create:
-                raise forms.ValidationError(self.clean_error_cantcreate())
-            else:
-                self.field_checker(
-                    str(self.instance.assigned_queue),
-                    str(self.cleaned_data.get('assigned_queue')))
-        return self.cleaned_data.get('assigned_queue')
-
-    def clean_assigned_user(self):
-        if self.instance.pk is not None:   # new instance only
-            self.field_checker(self.instance.assigned_user, self.cleaned_data.get('assigned_user'))
-        return self.cleaned_data.get('assigned_user')
-
-    def clean_subject(self):
-        if self.instance.pk is not None:  # new instance only
-            self.field_checker(self.instance.subject, self.cleaned_data.get('subject'))
-        return self.cleaned_data.get('subject')
-
-    def clean_body(self):
-        if self.instance.pk is not None:  # new instance only
-            self.field_checker(self.instance.body, self.cleaned_data.get('body'))
-        return self.cleaned_data.get('body')
-
-    '''
-    Error codes:
-    '''
-
-    def clean_error_cantedit(self):
-        cantedit = 'You don\'t have permissions to edit this ticket'
-        return cantedit
-
-    def clean_error_cantview(self):
-        cantview = 'You don\'t have permissions to view this ticket'
-        return cantview
-
-    def clean_error_cantcreate(self):
-        cantcreate = 'You don\'t have permissions to create this ticket'
-        return cantcreate
-
-    def clean_error_cantsave(self):
-        cantsave = 'You don\'t have permissions to save this ticket'
-        return cantsave
-
-    '''
-    Functions:
-    Check source=>destiny object and if it's not the same, log clean_it
-    '''
-
-    def field_checker(self, source=None, destiny=None, destiny_name=None):
-        if source != destiny:
-            logger(self.instance, self.request.user, "Changed", destiny)
-            pass
-
-    def clean(self):
-        # Some essential vars
-        user_obj = self.request.user
-        cleaned_data = self.cleaned_data
-        queue_obj = cleaned_data.get('assigned_queue')
-
-        '''Check creation'''
-        if not self.instance.pk:
-            user_object_rights = rights.get_rights_for_ticket(
-                user=user_obj, queue=queue_obj, ticket_id=None)
-            if not user_object_rights.can_create:
-                raise forms.ValidationError(self.clean_error_cantcreate())
-
-        '''Check edition'''
-        if self.instance.pk:
-            user_object_rights = rights.get_rights_for_ticket(
-                user=user_obj, queue=queue_obj, ticket_id=self.instance.id)
-            if not user_object_rights.can_edit:
-                raise forms.ValidationError(self.clean_error_cantedit())
-
-        '''Force to assign company'''
-        cleaned_data['assigned_company'] = queue_obj.company_rel
-
-        #  '''Put percentage 100% when closed ticket is detected'''
-        #  if cleaned_data['assigned_state'].id ==  3:
-        #      self.instance.percentage = int(100)
 
 
 # => Attachments
