+++ ticketatorapp/core/utils.py
