+++ ticketatorapp/core/views/group.py
@@ -1,7 +1,7 @@
 # Company views: list, create, delete
 
 from django.contrib.auth.models import Group
-from core.models import GroupForm
+from core.forms import GroupForm
 from django.shortcuts import render, get_object_or_404, redirect
 from django.http import Http404
