+++ ticketatorapp/core/views/inventory.py
